\name{geocrinia}
\docType{data}
\alias{geocrinia}
\title{Age data from populations of \emph{Geocrinia laevis}, \emph{G. victoriana} and their hybrids}
\description{Includes ages, sexes, populations and whether hybrid or not}
\usage{geocrinia}
\format{A dataframe containing ages of frogs (determined using skeletochronology), along with covariates for sex, population and whether hybrid or not}
\source{Scroggie, M.P. (2001) Zonal hybridization in the \emph{Geocrinia laevis} complex (Anura, Myobatrachidae): population ecology and male acoustic behaviour. PhD thesis, University of Melbourne}
