\name{geocrinia}
\docType{data}
\alias{geocrinia}
\title{Age data from populations of \emph{Geocrinia laevis}, \emph{G. victoriana} and their hybrids}
\description{Includes ages, sexes, populations and whether hybrid or not}
\usage{geocrinia}
\format{A dataframe containing ages of frogs (determined using skeletochronology), along with covariates for sex, population of origin, and whether hybrid or not}
\source{Scroggie, M.P. (2012) Survival of adult smooth froglets (\emph{Geocrinia laevis} complex, Anura, Myobatrachidae) in and around a hybrid zone. Herpetological Conservation and Biology 152: 136-144.
}
