\name{geocrinia}
\docType{data}
\alias{geocrinia}
\title{Age data from populations of Geocrinia laevis, G. victoriana and their hybrids}
\description{Includes ages, sexes, populations and whether hybrid or not.}
\usage{geocrinia}
 
\format{A dataframe containing ages of frogs, along with sexes, populations and whether hybrid or not}
\source{Scroggie, M.P. (2001) PhD thesis, University of Melbourne}
